Switch API benchmark to direct-JIT mode and remove bench_api_jit
